+++ CHANGELOG.md
@@ -5,6 +5,11 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+* Add labeling related code. Read the column data in `token` form or the `entire contents`.
+
 ## [0.7.0] - 2021-10-20
 
 ### Added